requirements.txt
absl-py==0.13.0
anyio==3.2.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chainer==7.8.0
chardet==4.0.0
cloudpickle==1.6.0
cntk-gpu==2.7
contextvars==2.4
cupy==9.1.0
cycler==0.10.0
Cython==0.29.23
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dm-sonnet==2.0.0
dm-tree==0.1.6
docopt==0.6.2
easydict==1.11
entrypoints==0.3
enum34==1.1.10
fastrlock==0.6
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.34.1
h5py==3.1.0
idna==2.10
imageio==2.9.0
immutables==0.15
importlib-metadata==4.5.0
install==1.3.5
iopath==0.1.10
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
jupyterlab-widgets==1.0.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
leveldb==0.1
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
msgpack==1.0.5
mxnet-cu110==1.8.0.post0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nose==1.3.7
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
onnx==1.9.0
onnxruntime==1.8.0
opencv-python==4.6.0.66
opencv-python-headless==4.2.0.34
opt-einsum==3.3.0
packaging==20.9
paddlepaddle-gpu==2.1.0
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipreqs==0.4.11
portalocker==2.7.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==4.21.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pycparser==2.20
Pygments==2.9.0
pygobject==3.26.1
pygpu==0.7.6
pyparsing==2.4.7
pyrsistent==0.17.3
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.8.1
python-distutils-extra==2.39
python-gflags==1.5.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.7.1
simplegeneric==0.8.1
six==1.15.0
sniffio==1.2.0
ssh-import-id==5.7
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.6.2.2
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
tensorflow-probability==0.13.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.10.0+cu111
torchvision==0.11.0+cu111
tornado==6.1
tqdm==4.61.1
traitlets==4.3.3
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yacs==0.1.8
yarg==0.1.9
zipp==3.4.1
